crud

+++ app/Http/Controllers/UserController.php
@@ -5,7 +5,11 @@ namespace App\Http\Controllers;
 use App\Http\Resources\UserReasource;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Redirect;
+use Illuminate\Support\Facades\Request as FacadesRequest;
+use Illuminate\Validation\Rule;
+use Illuminate\Validation\Rules\Password;
 use Inertia\Inertia;
 use Inertia\Response;
 
@@ -18,11 +22,9 @@ class UserController extends Controller
      */
     public function index()
     {
-
         return Inertia::render('Users/Index', [
             'items' => UserReasource::collection(User::paginate(50))
         ]);
-        
     }
 
     /**
@@ -32,7 +34,7 @@ class UserController extends Controller
      */
     public function create()
     {
-        //
+        return Inertia::render('Users/Create');
     }
 
     /**
@@ -43,7 +45,22 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        $request->validate([
+            'firstname' => 'required|string|max:255',
+            'lastname' => 'required|string|max:255',
+            'email' => 'required|string|email|max:255|unique:users',
+            'password' => ['required', Password::defaults()],
+        ]);
+
+        $user = User::create([
+            'firstname' => $request->firstname,
+            'lastname' => $request->lastname,
+            'email' => $request->email,
+            'phone' => $request->phone,
+            'password' => Hash::make($request->password),
+        ]);
+
+        return Redirect::route('users.index');
     }
 
     /**
@@ -63,9 +80,11 @@ class UserController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(User $user)
     {
-        //
+        return Inertia::render('Users/Edit', [
+            "user" => new UserReasource($user)
+        ]);
     }
 
     /**
@@ -77,7 +96,26 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+
+        $request->validate([
+            'firstname' => 'required|string|max:255',
+            'lastname' => 'required|string|max:255',
+            'email' => ['required', 'max:255', 'email', Rule::unique('users')->ignore($id)],
+            'phone' => ['nullable', 'max:25'],
+            'password' => ['nullable'],
+        ]);
+
+        $user = User::find($id);
+        if ($user) {
+            $user->update(FacadesRequest::only('firstname', 'lastname', 'email', 'phone'));
+
+            if ($request->get('password')) {
+                $user->update(['password' => Hash::make($request->get('password'))]);
+            }
+        }
+
+
+        return Redirect::route('users.index');
     }
 
     /**
@@ -88,9 +126,11 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $user=User::find($id);
-        $user->delete();
+        $user = User::find($id);
+        if ($user) {
+            $user->delete();
+        }
 
-        return Redirect::back()->with('success', 'User deleted.');
+        return Redirect::route('users.index');
     }
 }
